2:06 Added death

+++ Rogalic/Assets/Scripts/ManaBar.cs
@@ -12,13 +12,13 @@ public class ManaBar : MonoBehaviour
         slider = this.GetComponent<Slider>();
     }
 
-    public void SetMaxMana(int mana)
+    public void SetMaxMana(float mana)
     {
         slider.maxValue = mana;
         slider.value = mana;
     }
 
-    public void SetMana(int mana)
+    public void SetMana(float mana)
     {
         slider.value = mana;
     }
